+++ src/main/java/com/zenfulcode/commercify/commercify/dto/OrderLineDTO.java
@@ -9,9 +9,9 @@ import lombok.Data;
 @Builder
 @AllArgsConstructor
 public class OrderLineDTO {
-    private Long id;
-    private Long productId;
-    private Long variantId;
+    private Integer id;
+    private Integer productId;
+    private Integer variantId;
     private Integer quantity;
     private Double unitPrice;
     private String currency;
