+++ src/main/java/com/zenfulcode/commercify/commercify/exception/OrderNotFoundException.java
@@ -1,7 +1,7 @@
 package com.zenfulcode.commercify.commercify.exception;
 
 public class OrderNotFoundException extends RuntimeException {
-    public OrderNotFoundException(Long orderId) {
+    public OrderNotFoundException(Integer orderId) {
         super("Order not found with ID: " + orderId);
     }
 }
